// src/app/models/team.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Hackathon } from '../models/hackathon';
import { TeamMembers } from '../models/team-members';
export interface Team {
  createdAt?: string;
  hackathon?: Hackathon;
  id?: number;
  isFull?: boolean;
  isPublic?: boolean;
  joinCodeExpirationTime?: string;
  teamCode?: string;
  teamMembers?: Array<TeamMembers>;
  teamName?: string;
}