// src/app/models/team-members.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Team } from '../models/team';
import { User } from '../models/user';
export interface TeamMembers {
  id?: number;
  joinedAt?: string;
  role?: 'LEADER' | 'MEMBER' | 'MENTOR';
  team?: Team;
  user?: User;
}